Table Testing
Extensively tested to make sure the table was handling the data properly. Connected the table to the tbody code with D3, used Inspect to make sure of class ID tags for linking, and used console.log announcements to help in debugging when the table wasn't populating. Now just need to clean up the notes and the code and the comments.